chore(rxApp): decompose monolithic code
