+++ src/rxApp/scripts/rxAccountSearch.js
@@ -1,0 +1,20 @@
+angular.module('encore.ui.rxApp')
+/**
+ * @ngdoc directive
+ * @name rxApp.directive:rxAccountSearch
+ * @restrict E
+ * @description [TBD]
+ */
+.directive('rxAccountSearch', function ($window) {
+    return {
+        templateUrl: 'templates/rxAccountSearch.html',
+        restrict: 'E',
+        link: function (scope) {
+            scope.fetchAccount = function (searchValue) {
+                if (!_.isEmpty(searchValue)) {
+                    $window.location = '/search?term=' + searchValue;
+                }
+            };
+        }
+    };
+});
