+++ src/rxApp/scripts/rxApp.page.js
@@ -2,7 +2,6 @@
 var Page = require('astrolabe').Page;
 
 var rxApp = {
-
     cssCollapseButtonSelector: {
         /**
           Keep just the css string available for both the button element
@@ -93,7 +92,6 @@ var rxApp = {
             return this.exception('Navigation menu not collapsible');
         }
     }
-
 };
 
 exports.rxApp = {
@@ -103,64 +101,4 @@ exports.rxApp = {
         };
         return Page.create(rxApp);
     }
-};
-
-var rxPage = {
-
-    lblTitle: {
-        get: function () { return this.rootElement.$('.page-title > span'); }
-    },
-
-    lblSubtitle: {
-        get: function () { return this.rootElement.$('.page-subtitle'); }
-    },
-
-    lblTitleTag: {
-        get: function () { return this.rootElement.$('.page-titles .status-tag'); }
-    },
-
-    title: {
-        get: function () {
-            var page = this;
-            return this.lblTitle.isPresent().then(function (present) {
-                return present ? page.lblTitle.getText() : null;
-            });
-        }
-    },
-
-    subtitle: {
-        get: function () {
-            var page = this;
-            return this.lblSubtitle.isPresent().then(function (present) {
-                return present ? page.lblSubtitle.getText() : null;
-            });
-        }
-    },
-
-    titleTag: {
-        get: function () {
-            var page = this;
-            return this.lblTitleTag.isPresent().then(function (present) {
-                return present ? page.lblTitleTag.getText() : null;
-            });
-        }
-    }
-};
-
-exports.rxPage = {
-
-    initialize: function (rxPageElement) {
-        rxPage.rootElement = {
-            get: function () { return rxPageElement; }
-        };
-        return Page.create(rxPage);
-    },
-
-    main: (function () {
-        rxPage.rootElement = {
-            get: function () { return $('html'); }
-        };
-        return Page.create(rxPage);
-    })()
-
 };
