+++ src/rxApp/scripts/rxAppNavItem.js
@@ -1,0 +1,135 @@
+angular.module('encore.ui.rxApp')
+/**
+ * @ngdoc directive
+ * @name rxApp.directive:rxAppNavItem
+ * @restrict E
+ * @scope
+ * @description
+ * Creates a menu item. Recursively creates rx-app-nav if 'children' present.
+ * 'Item' must be avialable via scope
+ *
+ * @example
+ * <pre>
+ * <rx-app-nav-item ng-repeat="item in items"></rx-app-nav-item>
+ * </pre>
+ */
+.directive('rxAppNavItem', function ($compile, $location, $route) {
+    var linker = function (scope, element) {
+        var injectContent = function (selector, content) {
+            var el = element[0].querySelector(selector);
+            el = angular.element(el);
+
+            $compile(content)(scope, function (compiledHtml) {
+                el.append(compiledHtml);
+            });
+        };
+
+        var directiveHtml = '<directive></directive>';
+        // add navDirective if defined
+        if (angular.isString(scope.item.directive)) {
+            // convert directive string to HTML
+            // e.g. my-directive -> <my-directive></my-directive>
+            directiveHtml = directiveHtml.replace('directive', scope.item.directive);
+
+            injectContent('.item-directive', directiveHtml);
+        }
+
+        // increment nesting level for child items
+        var childLevel = scope.$parent.level + 1;
+        // safety check that child level is a number
+        if (isNaN(childLevel)) {
+            childLevel = 2;
+        }
+        // add children if present
+        // Note: this can't be added in the HTML due to angular recursion issues
+        var rxNavTemplate = '<rx-app-nav items="item.children" level="' + childLevel + '">' +
+            '</rx-app-nav>';
+        if (angular.isArray(scope.item.children)) {
+            injectContent('.item-children', rxNavTemplate);
+        }
+    };
+
+    return {
+        restrict: 'E',
+        replace: true,
+        templateUrl: 'templates/rxAppNavItem.html',
+        link: linker,
+        scope: {
+            item: '='
+        },
+        controller: function ($scope, $location, rxVisibility, Permission) {
+            // provide `route` as a scope property so that links can tie into them
+            $scope.route = $route;
+
+            var roleCheck = function (roles) {
+                if (_.isUndefined(roles)) {
+                    return true;
+                }
+
+                if (!_.isUndefined(roles.any)) {
+                    return Permission.hasRole(roles.any);
+                }
+
+                if (!_.isUndefined(roles.all)) {
+                    return Permission.hasAllRoles(roles.all);
+                }
+
+                return false;
+            };
+
+            /*
+             * @description Determines whether or not a nav item should be displayed, based on `visibility`
+             * criteria and `roles` criteria
+             * @param [visibility] - Can be an expression, a function, an array (using format below) to
+             *                     determine visibility
+             * @param {object} [roles] - An object with a format { 'any': ['role1', 'role2'] } or
+             *                           { 'all': ['role1', 'role2'] }
+             */
+            $scope.isVisible = function (visibility, roles) {
+                var locals = {
+                    location: $location
+                };
+                if (_.isUndefined(visibility) && _.isUndefined(roles)) {
+                    // no visibility or role criteria specified, so default to true
+                    return true;
+                }
+
+                if (_.isArray(visibility)) {
+                    // Expected format is
+                    // ["someMethodName", { param1: "abc", param2: "def" }]
+                    // The second element of the array is optional, used to pass extra
+                    // info to "someMethodName"
+                    var methodName = visibility[0];
+                    var configObj = visibility[1]; //optional
+
+                    _.merge(locals, configObj);
+
+                    // The string 'false' will evaluate to the "real" false
+                    // in $scope.$eval
+                    visibility = rxVisibility.getMethod(methodName) || 'false';
+                }
+
+                // If `visibility` isn't defined, then default it to `true` (i.e. visible)
+                var visible = _.isUndefined(visibility) ? true : $scope.$eval(visibility, locals),
+                    hasRole = true;
+
+                // Only do a roleCheck() if `visible` is true. If we failed the visibility test,
+                // then we must ensure the nav item is not displayed, regardless of the roles
+                if (visible && _.isObject(roles)) {
+                    hasRole = roleCheck(roles);
+                }
+
+                return visible && hasRole;
+            };
+
+            $scope.toggleNav = function (ev, href) {
+                // if no href present, simply toggle active state
+                if (_.isEmpty(href)) {
+                    ev.preventDefault();
+                    $scope.item.active = !$scope.item.active;
+                }
+                // otherwise, let the default nav do it's thing
+            };
+        }
+    };
+});
