+++ src/rxApp/scripts/rxAppSearch.js
@@ -1,0 +1,26 @@
+angular.module('encore.ui.rxApp')
+/**
+ * @ngdoc directive
+ * @name rxApp.directive:rxAppSearch
+ * @restrict E
+ * @scope
+ * @description
+ * Creates a search input form for navigation
+ *
+ * @param {string} [placeholder] Title of page
+ * @param {*} [model] Model to tie input form to (via ng-model)
+ * @param {function} [submit] Function to run on submit (model is passed as only argument to function)
+ */
+.directive('rxAppSearch', function () {
+    return {
+        restrict: 'E',
+        replace: true,
+        templateUrl: 'templates/rxAppSearch.html',
+        scope: {
+            placeholder: '@?',
+            model: '=?',
+            submit: '=?',
+            pattern: '@?'
+        }
+    };
+});
