+++ src/rxApp/scripts/rxAtlasSearch.js
@@ -1,0 +1,21 @@
+angular.module('encore.ui.rxApp')
+/**
+ * @ngdoc directive
+ * @name rxApp.directive:rxAtlasSearch
+ * @restrict E
+ * @description
+ * Used to search accounts for Cloud Atlas
+ */
+.directive('rxAtlasSearch', function ($window) {
+    return {
+        template: '<rx-app-search placeholder="Search by username..." submit="searchAccounts"></rx-app-search>',
+        restrict: 'E',
+        link: function (scope) {
+            scope.searchAccounts = function (searchValue) {
+                if (!_.isEmpty(searchValue)) {
+                    $window.location = '/cloud/' + searchValue + '/servers/';
+                }
+            };
+        }
+    };
+});
