+++ src/rxApp/scripts/rxTicketSearch.js
@@ -1,0 +1,20 @@
+angular.module('encore.ui.rxApp')
+/**
+ * @ngdoc directive
+ * @name rxApp.directive:rxTicketSearch
+ * @restrict E
+ * @description
+ * Used to search tickets for Ticket Queues
+ */
+.directive('rxTicketSearch', function () {
+    return {
+        template: '<rx-app-search placeholder="Search for a Ticket..." submit="searchTickets"></rx-app-search>',
+        restrict: 'E',
+        link: function (scope) {
+            // TQTicketSelection.loadTicket.bind(TQTicketSelection)
+            scope.searchTickets = function () {
+                // TODO do something here
+            };
+        }
+    };
+});
