+++ src/rxApp/scripts/rxVisibilityPathParams.js
@@ -1,0 +1,22 @@
+angular.module('encore.ui.rxApp')
+/**
+ * @ngdoc service
+ * @name rxApp.service:rxVisibilityPathParams
+ * @description
+ * Returns an object with `name` and `method` params that can
+ * be passed to `rxVisibility.addMethod()`. We use register this by
+ * default, as it's used by the nav menu we keep in routesCdnPath.
+ * The method is used to check if {param: 'someParamName'} is present
+ * in the current route
+ * Use it as `visibility: [ 'rxPathParams', { param: 'userName' } ]`
+ */
+.factory('rxVisibilityPathParams', function ($routeParams) {
+    var pathParams = {
+        name:'rxPathParams',
+        method: function (scope, args) {
+            return !_.isUndefined($routeParams[args.param]);
+        }
+    };
+
+    return pathParams;
+});
